# apps/web/src/generated/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reframe V7 API",
    "description": "Dual-Run Retrospective Deep Narrative Forensics API with ClickHouse & Google ADK",
    "version": "7.0.0"
  },
  "paths": {
...
    "/api/v1/reframe-runs": {
      "post": {
        "tags": [
          "Reframe Runs"
        ],
        "summary": "Create Reframe Run",
        "description": "Submits a durable reframe narrative analysis run.\nStrictly forbids client spoiler_cutoff_ms or SQL tampering.",
        "operationId": "create_reframe_run_api_v1_reframe_runs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReframeRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reframe-runs/{run_id}": {
      "get": {
        "tags": [
          "Reframe Runs"
        ],
        "summary": "Get Reframe Run",
        "operationId": "get_reframe_run_api_v1_reframe_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Reframe Run Api V1 Reframe Runs  Run Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reframe-runs/{run_id}/events": {
      "get": {
        "tags": [
          "Reframe Runs"
        ],
        "summary": "Stream Reframe Run Events",
        "operationId": "stream_reframe_run_events_api_v1_reframe_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/audience-lab/runs/{run_id}": {
      "get": {
        "tags": [
          "Admin Audience Lab"
        ],
        "summary": "Get Audience Simulation Run",
        "description": "Fetches status and results of an audience simulation run.",
        "operationId": "get_audience_simulation_run_api_v1_admin_audience_lab_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateReframeRunRequest": {
        "properties": {
          "movie_id": {
            "type": "string",
            "title": "Movie Id",
            "default": "the-bat-whispers-1930"
          },
          "edition_id": {
            "type": "string",
            "title": "Edition Id",
            "default": "tbw-fullscreen-archive"
          },
          "reveal_id": {
            "type": "string",
            "title": "Reveal Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "STRICT_CANON"
          },
          "top_k": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "reveal_id"
        ],
        "title": "CreateReframeRunRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}